requirements.txt
numpy>=1.20
pandas>=1.3
scikit-learn>=1.0
spacy>=3.0
PyPDF2>=3.0
python-docx>=0.8
